Add icon_class window option and incrementing icons per instance.
Quick and dirty hack to allow to give an "icon_class" to alacritty. It
will then seatch for files "alacritty_icon_${icon_class}_[0-9]+.ico in
the same directory as the executable and use those icons. Every instance
that gets started uses the next of the icons of that icon_class. It
keeps track of the last used one in a file alacritty_icons.txt in
${TEMP}.

Note, to specify the icon class on the commandline, quotes are required:

alacritty.exe --option window.icon_class=\"what\"